refactor: Extract valve control helpers and fix safety override
- Add _is_boost_heating_active() to check boost condition
- Add _get_valve_control() to determine valve settings from boost or calibration
- Add _apply_valve_control() to apply valve settings to TRV
- Add _reset_valve_on_safety_override() to reset valve when HVAC forced OFF
- Remove redundant cooler section from unavailable TRV path (handled by control_cooler)
- Reduce code duplication between unavailable and available TRV paths
- Add tests for safety override scenarios (window open, no heat call)
- Rename test file from test_boost_mode_bug.py to test_boost_mode.py
